+++ BugTracker/BugTracker.sln
@@ -1,0 +1,25 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio Version 17
+VisualStudioVersion = 17.2.32630.192
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "BugTracker", "BugTracker.csproj", "{85E3FA46-5FAF-4664-A3F8-77DD7A246677}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{85E3FA46-5FAF-4664-A3F8-77DD7A246677}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{85E3FA46-5FAF-4664-A3F8-77DD7A246677}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{85E3FA46-5FAF-4664-A3F8-77DD7A246677}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{85E3FA46-5FAF-4664-A3F8-77DD7A246677}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {A980DC99-CDEF-462D-BD72-5FA228DB0EBE}
+	EndGlobalSection
+EndGlobal
