--- src/BugTracker/BugTracker.sln
@@ -1,25 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 16
-VisualStudioVersion = 16.0.31515.178
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "BugTracker", "BugTracker\BugTracker.csproj", "{3FBE4C8A-1F25-4DB9-8948-D65D9B99D541}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{3FBE4C8A-1F25-4DB9-8948-D65D9B99D541}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{3FBE4C8A-1F25-4DB9-8948-D65D9B99D541}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{3FBE4C8A-1F25-4DB9-8948-D65D9B99D541}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{3FBE4C8A-1F25-4DB9-8948-D65D9B99D541}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {A2DCCDB9-136A-4705-B356-B93457AE268F}
-	EndGlobalSection
-EndGlobal
