--- src/BugTracker/BugTracker/frmLogin.cs
@@ -1,65 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.ComponentModel;
-using System.Data;
-using System.Drawing;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-using System.Windows.Forms;
-
-namespace BugTracker
-{
-    public partial class frmLogin : Form
-    {
-        public frmLogin()
-        {
-            InitializeComponent();
-        }
-
-        private void btnSalir_Click(object sender, EventArgs e)
-        {
-            Environment.Exit(0);
-        }
-
-        private void btnIngresar_Click(object sender, EventArgs e)
-        {
-            if (txtUsuario.Text == "")
-            {
-                MessageBox.Show("Se debe ingresar un usuario.", "Aviso", MessageBoxButtons.OK, MessageBoxIcon.Exclamation);
-                return;
-            }
-
-            if (txtPassword.Text == "")
-            {
-                MessageBox.Show("Se debe ingresar una contraseña.", "Aviso", MessageBoxButtons.OK, MessageBoxIcon.Exclamation);
-                return;
-            }
-
-
-            if (ValidarCredenciales(txtUsuario.Text, txtPassword.Text))
-            {
-                MessageBox.Show("Usted a ingresado al sistema.", "Aviso", MessageBoxButtons.OK, MessageBoxIcon.Information);
-                
-                this.Close();
-            }
-            else
-            {
-                txtUsuario.Text = "";
-                txtPassword.Text = "";
-                MessageBox.Show("Debe ingresar usuario y/o contraseña válidos", "Aviso", MessageBoxButtons.OK, MessageBoxIcon.Error);
-            }
-        }
-
-        public bool ValidarCredenciales(string pUsuario, string pPassword)
-        {
-            return true;
-        }
-
-        private void frmLogin_Load(object sender, EventArgs e)
-        {
-            this.CenterToParent();
-        }
-    }
-
-}
